requirements.txt
# Smart NOTAM - 대한항공 NOTAM 처리 시스템
# 필수 패키지 목록

# 웹 프레임워크
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0

# PDF 처리
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image

# AI/번역
openai==1.3.8
google-generativeai==0.3.2
requests==2.31.0

# 데이터 처리
pandas==2.1.3
numpy==1.25.2
xarray
herbie-data
cfgrib
scipy

# 텍스트 처리
regex==2023.10.3

# 환경 설정
python-dotenv==1.0.0

# 로깅
colorlog==6.7.0

# 시간 처리
python-dateutil==2.8.2

# JSON 처리
jsonschema==4.19.2

# HTTP 클라이언트
urllib3==2.0.7

# 보안
cryptography==41.0.7

# 개발용 도구 (선택사항)
pytest==7.4.3
pytest-flask==1.3.0
black==23.10.1
flake8==6.1.0

# 추가 유틸리티
click==8.1.7
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2